// src/components/main-footer/main-footer.types.d.ts
export interface Settings {
    menu_1_title: string;
    menu_1: string;
    menu_2_title: string;
    menu_2: string;
    menu_3_title: string;
    menu_3: string;
    logo: string | null;
    description: string;
    email: string;
    auxiliar: string;
    address: string;
    copyright: string;
}